requirements.txt
# ============================================================================
# Core Django Framework
# ============================================================================
Django==5.0.6

# ============================================================================
# Celery and Related Packages
# ============================================================================
celery==5.4.0                  # Distributed task queue
redis==5.0.1                   # Redis client for Python
django-celery-results==2.5.1   # Store Celery task results in Django database
django-celery-beat==2.6.0      # Periodic task scheduler
flower==2.0.1                  # Real-time Celery monitoring tool

# ============================================================================
# Cache Backend
# ============================================================================
django-redis==5.4.0            # Redis cache backend for Django

# ============================================================================
# Environment Variables
# ============================================================================
python-decouple==3.8           # Separate settings from code

# ============================================================================
# Database Drivers (Optional - Uncomment as needed)
# ============================================================================
# PostgreSQL
# psycopg2-binary==2.9.9

# MySQL
# mysqlclient==2.2.4

# ============================================================================
# API and Serialization
# ============================================================================
djangorestframework==3.14.0    # REST API framework (optional but recommended)

# ============================================================================
# Development and Testing
# ============================================================================
# Testing
pytest==8.0.0
pytest-django==4.8.0
pytest-celery==0.0.0
coverage==7.4.1

# Code Quality
flake8==7.0.0
black==24.2.0
isort==5.13.2

# Debugging
django-debug-toolbar==4.3.0
ipython==8.21.0

# ============================================================================
# Production Server (Optional)
# ============================================================================
gunicorn==21.2.0              # WSGI HTTP Server
whitenoise==6.6.0             # Static file serving

# ============================================================================
# Monitoring and Logging (Optional)
# ============================================================================
sentry-sdk==1.40.0            # Error tracking
prometheus-client==0.19.0      # Metrics

# ============================================================================
# Utilities
# ============================================================================
python-dateutil==2.8.2
pytz==2024.1
